skipping existing files

+++ main.py
@@ -7,11 +7,12 @@ import ffmpeg
 from extract_features import run
 from resnet import i3_res50
 from tqdm import tqdm
+import os.path
 
-
-def generate(datasetpath, outputpath, pretrainedpath, frequency, batch_size, sample_mode, use_cuda, vid_ext):
+def generate(datasetpath, outputpath, pretrainedpath, frequency, batch_size, sample_mode, use_cuda, vid_ext, overwrite):
 	Path(outputpath).mkdir(parents=True, exist_ok=True)
-	temppath = outputpath+ "/temp/"
+	temppath = outputpath + "/temp/"
+	shutil.rmtree(temppath) # make sure temp dir is empty, can be leftovers
 	rootdir = Path(datasetpath)
 	videos = [str(f) for f in rootdir.glob('**/*.' + vid_ext)]
 	# setup the model
@@ -21,13 +22,16 @@ def generate(datasetpath, outputpath, pretrainedpath, frequency, batch_size, sam
 	i3d.train(False)  # Set model to evaluate mode
 	for video in tqdm(videos):
 		videoname = video.split("/")[-1].split(".")[0]
+		features_output_fn = outputpath + "/" + videoname
+		if not overwrite and os.path.isfile(features_output_fn):
+			continue # skip extracted features
 		startime = time.time()
 		print("Generating JPG files for each frame of {0}...".format(video))
 		Path(temppath).mkdir(parents=True, exist_ok=True)
 		ffmpeg.input(video).output('{}%d.jpg'.format(temppath),start_number=0).global_args('-loglevel', 'quiet').run()
 		print("Extracting features from frames...")
 		features = run(i3d, frequency, temppath, batch_size, sample_mode, use_cuda)
-		np.save(outputpath + "/" + videoname, features)
+		np.save(features_output_fn, features)
 		print("Obtained features of shape", features.shape, "processing time {0} seconds".format(time.time() - startime))
 		shutil.rmtree(temppath)
 
@@ -41,5 +45,6 @@ if __name__ == '__main__':
 	parser.add_argument('--sample_mode', type=str, default="oversample")
 	parser.add_argument('--ext', type=str, default="mkv")
 	parser.add_argument('--cuda', action='store_true')
+	parser.add_argument('--overwrite', action='store_true')
 	args = parser.parse_args()
-	generate(args.datasetpath, str(args.outputpath), args.pretrainedpath, args.frequency, args.batch_size, args.sample_mode, args.cuda, args.ext)    
+	generate(args.datasetpath, str(args.outputpath), args.pretrainedpath, args.frequency, args.batch_size, args.sample_mode, args.cuda, args.ext, args.overwrite)
